+++ src/core/graphql/scalars/index.ts
@@ -1,0 +1,3 @@
+export * from './email.scalar'
+export * from './url.scalar'
+export * from './uuid.scalar'