+++ src/core/proto/auth-sign-up.ts
@@ -1,3 +1,0 @@
-import { UserCreate } from '~/core/proto/user-create'
-
-export class AuthSignUp extends UserCreate {}