+++ CHANGELOG.md
@@ -9,13 +9,13 @@ The format is based on [Keep a Changelog].
 - `CHANGELOG.md` file ([3])
 - Swagger UI for API documentation ([1])
 - PostgreSQL database integration via Drizzle ORM ([2])
-- Set up `UserModule` and `UserService` with basic `create` method ([23])
 - Set up `AuthModule` and primitive authentication functionality ([25])
 - Set up `ProjectModule` with basic create/read functionality ([29])
 - Extra API model `PaginationQuery` ([29])
 - Set up `TokenModule` with basic create/read functionality ([34])
 - Set up `AccountModule` and `AccountService` with basic functionality ([49])
 - GitHub OAuth2 integration ([44])
+- GraphQL API integration ([54])
 
 ### Changed
 
@@ -40,6 +40,8 @@ The format is based on [Keep a Changelog].
 - Added `@fastify/static`, `@nestjs/platform-fastify` packages ([52])
 - Migrated to Fastify as the HTTP provider ([52])
 - Removed `@nestjs/platform-express`, `@types/express` ([54])
+- Added `@nestjs/graphql`, `graphql` packages ([54])
+- Added `@nestjs/mercurius`, `mercurius` packages ([54])
 
 <!-- Links -->